requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.2.1
django-allauth==0.61.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.2.0
idna==3.10
mccabe==0.7.0
oauthlib==3.2.2
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
